refactor(Cypress): Splitting commands.ts
support/commands.ts - split into multiple files in a new commands
dir that group the commands based on their purpose.

support/nav.ts - integrated under the commands/utils.ts command
